+++ src/components/Autocomplete/Autocomplete.tsx
@@ -5,12 +5,13 @@ import TextField from "components/TextField";
 import type { TextFieldProps } from "components/TextField";
 import DropdownMenu from "components/DropdownMenu";
 import type { MenuItemProps } from "components/MenuItem";
-import { getActiveElement } from "utilities/a11y";
 import * as keys from "constants/keys";
 import useHotkeys from "hooks/useHotkeys";
 import useHandlerRef from "hooks/useHandlerRef";
 import * as T from "./Autocomplete.types";
 import s from "./Autocomplete.module.css";
+import useElementId from "hooks/useElementId";
+import useIsomorphicLayoutEffect from "hooks/useIsomorphicLayoutEffect";
 
 const AutocompleteContext = React.createContext({} as T.Context);
 
@@ -30,8 +31,9 @@ const Autocomplete: React.FC<T.Props> & { Item: typeof AutocompleteItem } = (pro
 		onClose,
 		...textFieldProps
 	} = props;
+	const [highlightedId, setHighlightedId] = React.useState<string>();
 	const onBackspaceRef = useHandlerRef(onBackspace);
-	const onEnterRef = useHandlerRef(onEnter);
+	const contentRef = React.useRef<HTMLDivElement>(null);
 	const internalInputRef = React.useRef<HTMLInputElement>(null);
 	const inputAttributesRef = textFieldProps.inputAttributes?.ref;
 	const inputRef =
@@ -43,16 +45,19 @@ const Autocomplete: React.FC<T.Props> & { Item: typeof AutocompleteItem } = (pro
 	const lockedRef = React.useRef(false);
 	const onOpenRef = useHandlerRef(onOpen);
 	const onCloseRef = useHandlerRef(onClose);
+	const onChangeRef = useHandlerRef(onChange);
+	const onItemSelectRef = useHandlerRef(onItemSelect);
+	const onEnterRef = useHandlerRef(onEnter);
 	const isDropdownActive = hasChildren && (active ?? internalActive);
 
-	const lockDropdown = () => {
+	const lockDropdown = React.useCallback(() => {
 		// Prevent dropdown from re-opening when clicked on item with mouse
 		// and focus moves to the item and back to the input
 		lockedRef.current = true;
 		setTimeout(() => {
 			lockedRef.current = false;
 		}, 100);
-	};
+	}, []);
 
 	const handleOpen = React.useCallback(() => {
 		if (lockedRef.current) return;
@@ -65,11 +70,14 @@ const Autocomplete: React.FC<T.Props> & { Item: typeof AutocompleteItem } = (pro
 		onCloseRef.current?.(args);
 	};
 
-	const handleItemClick: T.Context["onItemClick"] = (args) => {
-		onChange?.({ value: args.value, name });
-		onItemSelect?.(args);
-		lockDropdown();
-	};
+	const handleItemClick: T.Context["onItemClick"] = React.useCallback(
+		(args) => {
+			onChangeRef.current?.({ value: args.value, name });
+			onItemSelectRef.current?.(args);
+			lockDropdown();
+		},
+		[lockDropdown, onChangeRef, onItemSelectRef, name]
+	);
 
 	const handleChange: TextFieldProps["onChange"] = (args) => {
 		onChange?.(args);
@@ -91,43 +99,79 @@ const Autocomplete: React.FC<T.Props> & { Item: typeof AutocompleteItem } = (pro
 		inputRef.current?.focus();
 	};
 
+	const getOptionElements = React.useCallback(() => {
+		const contentEl = contentRef.current;
+
+		if (!contentEl) return [];
+		return Array.from(contentEl.querySelectorAll("[role=option]")) as HTMLElement[];
+	}, []);
+
 	useHotkeys(
 		{
+			[keys.ENTER]: () => {
+				const options = getOptionElements();
+				const highlightedOption = options.find((el) => el.id === highlightedId);
+
+				highlightedOption?.click();
+				onEnterRef.current?.();
+			},
 			[keys.BACKSPACE]: () => {
 				onBackspaceRef.current?.();
 			},
-			[keys.ENTER]: () => {
-				onEnterRef.current?.();
-			},
-		},
-		[onBackspaceRef, onEnterRef],
-		{
-			ref: inputRef,
-			disabled: !onBackspaceRef.current && !onEnterRef.current,
-		}
-	);
+			[keys.UP]: () => {
+				const options = getOptionElements();
 
-	useHotkeys(
-		{
+				if (options.length) {
+					const highlightedIndex = options.findIndex((el) => el.id === highlightedId);
+					const nextOption = options.at(highlightedIndex - 1) || options.at(-1)!;
+
+					setHighlightedId(nextOption.id);
+				}
+			},
 			[keys.DOWN]: () => {
 				handleOpen();
-			},
-			[keys.ENTER]: () => {
-				const el = getActiveElement(inputRef.current);
-				el?.click();
+
+				const options = getOptionElements();
+
+				if (options.length) {
+					const highlightedIndex = options.findIndex((el) => el.id === highlightedId);
+					const nextOption = options.at(highlightedIndex + 1) || options.at(0)!;
+
+					setHighlightedId(nextOption.id);
+				}
 			},
 		},
-		[handleOpen],
+		[handleOpen, getOptionElements, highlightedId],
 		{ ref: inputRef, preventDefault: true }
 	);
 
+	useIsomorphicLayoutEffect(() => {
+		if (!isDropdownActive) return;
+
+		requestAnimationFrame(() => {
+			const options = getOptionElements();
+			const firstId = options[0]?.id;
+
+			if (firstId) setHighlightedId(firstId);
+		});
+	}, [isDropdownActive]);
+
+	const contextValue = React.useMemo(
+		() => ({
+			onItemClick: handleItemClick,
+			highlightedId,
+			setHighlightedId,
+		}),
+		[highlightedId, handleItemClick]
+	);
+
 	return (
-		<AutocompleteContext.Provider value={{ onItemClick: handleItemClick }}>
+		<AutocompleteContext.Provider value={contextValue}>
 			<DropdownMenu
 				position="bottom"
 				width="trigger"
 				triggerType="focus"
-				trapFocusMode="selection-menu"
+				trapFocusMode={false}
 				active={isDropdownActive}
 				onClose={handleClose}
 				onOpen={handleOpen}
@@ -162,11 +206,16 @@ const Autocomplete: React.FC<T.Props> & { Item: typeof AutocompleteItem } = (pro
 								onClick: attributes.onFocus,
 								ref: inputRef,
 								role: "combobox",
+								"aria-activedescendant": highlightedId,
+								"aria-haspopup": "listbox",
+								"aria-autocomplete": "list",
 							}}
 						/>
 					)}
 				</DropdownMenu.Trigger>
-				<DropdownMenu.Content attributes={{ onClick: handleContentClick }}>
+				<DropdownMenu.Content
+					attributes={{ onClick: handleContentClick, role: "listbox", ref: contentRef }}
+				>
 					{children}
 				</DropdownMenu.Content>
 			</DropdownMenu>
@@ -176,7 +225,8 @@ const Autocomplete: React.FC<T.Props> & { Item: typeof AutocompleteItem } = (pro
 
 const AutocompleteItem: React.FC<T.ItemProps> = (props) => {
 	const { value, data, onClick, ...menuItemProps } = props;
-	const { onItemClick } = React.useContext(AutocompleteContext);
+	const { onItemClick, highlightedId } = React.useContext(AutocompleteContext);
+	const id = useElementId();
 
 	const handleClick: MenuItemProps["onClick"] = (e) => {
 		onClick?.(e);
@@ -187,9 +237,12 @@ const AutocompleteItem: React.FC<T.ItemProps> = (props) => {
 		<DropdownMenu.Item
 			{...menuItemProps}
 			className={[menuItemProps.disabled && s["item--disabled"], menuItemProps.className]}
+			highlighted={highlightedId === id}
 			attributes={{
 				...menuItemProps.attributes,
 				role: "option",
+				id,
+				tabIndex: -1,
 			}}
 			onClick={handleClick}
 		/>